requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==3.2.*
django-cache-url==3.3.0
django-parler==2.2.1
django-rosetta==0.9.3
environs==9.3.5
gunicorn==20.1.0
idna==3.3
marshmallow==3.14.1
Pillow==9.0.1
polib==1.1.1
psycopg2==2.9.*
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.8
